Wrangle SBV backend into 'working' order.
Basically connect all the hoses & use the old backend to fill in the remaining
missing gap that is needed to generate canonical forms for polymorphic
unit-vars. Otherwise, this uses SBV/Z3 to solve all of the monomorphic parts of
the problem. Which, as it turns out, is usually most of the problem.